+++ be/src/exec/pipeline/exchange/multi_cast_local_exchange.h
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#pragma once
+
 #include <deque>
 #include <utility>
 
@@ -194,105 +196,6 @@ private:
     // Cell* _flush_point = nullptr;
     mutable Cell* _flush_point = nullptr;
     std::shared_ptr<MemLimitedChunkQueue> _queue;
-};
-
-
-
-// ===== source op =====
-class MultiCastLocalExchangeSourceOperator final : public SourceOperator {
-public:
-    MultiCastLocalExchangeSourceOperator(OperatorFactory* factory, int32_t id, int32_t plan_node_id,
-                                         int32_t driver_sequence, int32_t mcast_consumer_index,
-                                         std::shared_ptr<MultiCastLocalExchanger> exchanger)
-            : SourceOperator(factory, id, "multi_cast_local_exchange_source", plan_node_id, true, driver_sequence),
-              _mcast_consumer_index(mcast_consumer_index),
-              _exchanger(std::move(std::move(exchanger))) {}
-
-    Status prepare(RuntimeState* state) override;
-
-    bool has_output() const override;
-
-    bool is_finished() const override { return _is_finished; }
-
-    Status set_finishing(RuntimeState* state) override;
-
-    StatusOr<ChunkPtr> pull_chunk(RuntimeState* state) override;
-
-private:
-    bool _is_finished = false;
-    int32_t _mcast_consumer_index;
-    std::shared_ptr<MultiCastLocalExchanger> _exchanger;
-};
-
-class MultiCastLocalExchangeSourceOperatorFactory final : public SourceOperatorFactory {
-public:
-    MultiCastLocalExchangeSourceOperatorFactory(int32_t id, int32_t plan_node_id, int32_t mcast_consumer_index,
-                                                std::shared_ptr<MultiCastLocalExchanger> exchanger)
-            : SourceOperatorFactory(id, "multi_cast_local_exchange_source", plan_node_id),
-              _mcast_consumer_index(mcast_consumer_index),
-              _exchanger(std::move(std::move(exchanger))) {}
-    ~MultiCastLocalExchangeSourceOperatorFactory() override = default;
-    OperatorPtr create(int32_t degree_of_parallelism, int32_t driver_sequence) override {
-        return std::make_shared<MultiCastLocalExchangeSourceOperator>(this, _id, _plan_node_id, driver_sequence,
-                                                                      _mcast_consumer_index, _exchanger);
-    }
-
-private:
-    int32_t _mcast_consumer_index;
-    std::shared_ptr<MultiCastLocalExchanger> _exchanger;
-};
-
-// ===== sink op =====
-
-class MultiCastLocalExchangeSinkOperator final : public Operator {
-public:
-    MultiCastLocalExchangeSinkOperator(OperatorFactory* factory, int32_t id, int32_t plan_node_id,
-                                       const int32_t driver_sequence,
-                                       std::shared_ptr<MultiCastLocalExchanger> exchanger)
-            : Operator(factory, id, "multi_cast_local_exchange_sink", plan_node_id, true, driver_sequence),
-              _exchanger(std::move(std::move(exchanger))) {}
-
-    ~MultiCastLocalExchangeSinkOperator() override = default;
-
-    Status prepare(RuntimeState* state) override;
-
-    bool has_output() const override { return false; }
-
-    bool need_input() const override;
-
-    bool is_finished() const override { return _is_finished; }
-
-    Status set_finishing(RuntimeState* state) override;
-
-    StatusOr<ChunkPtr> pull_chunk(RuntimeState* state) override;
-
-    Status push_chunk(RuntimeState* state, const ChunkPtr& chunk) override;
-
-    void update_counter(size_t memory_usage, size_t buffer_row_size);
-
-private:
-    bool _is_finished = false;
-    const std::shared_ptr<MultiCastLocalExchanger> _exchanger;
-    RuntimeProfile::HighWaterMarkCounter* _peak_memory_usage_counter = nullptr;
-    RuntimeProfile::HighWaterMarkCounter* _peak_buffer_row_size_counter = nullptr;
-};
-
-class MultiCastLocalExchangeSinkOperatorFactory final : public OperatorFactory {
-public:
-    MultiCastLocalExchangeSinkOperatorFactory(int32_t id, int32_t plan_node_id,
-                                              std::shared_ptr<MultiCastLocalExchanger> exchanger)
-            : OperatorFactory(id, "multi_cast_local_exchange_sink", plan_node_id),
-              _exchanger(std::move(std::move(exchanger))) {}
-
-    ~MultiCastLocalExchangeSinkOperatorFactory() override = default;
-
-    OperatorPtr create(int32_t degree_of_parallelism, int32_t driver_sequence) override {
-        return std::make_shared<MultiCastLocalExchangeSinkOperator>(this, _id, _plan_node_id, driver_sequence,
-                                                                    _exchanger);
-    }
-
-private:
-    std::shared_ptr<MultiCastLocalExchanger> _exchanger;
 };
 
 } // namespace starrocks::pipeline
